Alter MSSQL Colstore Index Steps
Alter the Steps involved in creating colstore indices. Testing the
current process in loading data with Colstore indices created before
the data is loaded vs after the data is loaded shows a 10-20x
improvement when the index is created after loading the data, when BCP
is not used. When BCP is used it is more efficient to create the indices
before the data is loaded.